% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dna_extract_tail_boundaries_in_fastq_base_number.R
\name{dna_extract_tail_boundaries_in_fastq_base_number}
\alias{dna_extract_tail_boundaries_in_fastq_base_number}
\title{Title}
\usage{
dna_extract_tail_boundaries_in_fastq_base_number(event_data, start, end, fastq)
}
\arguments{
\item{fastq}{}
}
\value{

}
\description{
Title
}
