% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna-tailtype-finder.R
\name{dna_tailtype_finder}
\alias{dna_tailtype_finder}
\title{Title}
\usage{
dna_tailtype_finder(file_path = NA, data = "cdna",
  read_id_fast5_file = NA, plot_debug = F, multifast5 = F,
  basecalled_with_flipflop = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
