% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-read-data-hdf5r-flipflop-multifast5.R
\name{extract_read_data_hdf5r_flipflop_multifast5}
\alias{extract_read_data_hdf5r_flipflop_multifast5}
\title{Title}
\usage{
extract_read_data_hdf5r_flipflop_multifast5(read_id_fast5_file,
  plot_debug = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
