% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-read-data-rhdf5.R
\name{extract_read_data_rhdf5}
\alias{extract_read_data_rhdf5}
\title{Extract data from ONT Fast5 file}
\usage{
extract_read_data_rhdf5(read_path)
}
\arguments{
\item{read_path}{Full path of a FAST5 read}
}
\value{
A list of relevant data extracted from the FAST5 file
}
\description{
Extracts read data from basecalled FAST5 files. The function can't work on
raw reads that haven't been basecalled by Albacore.
}
\examples{
extract_read_data_rhdf5('path/to/fast5/file')
}
