% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tail-boundaries-in-fastq-base-number.R
\name{extract_tail_boundaries_in_fastq_base_number}
\alias{extract_tail_boundaries_in_fastq_base_number}
\title{Title}
\usage{
extract_tail_boundaries_in_fastq_base_number(event_data, start, end, fastq)
}
\arguments{
\item{fastq}{}
}
\value{

}
\description{
Title
}
