% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-cdna-polya.R
\name{find_cdna_polya}
\alias{find_cdna_polya}
\title{Find Poly(A) tail in a cDNA read}
\usage{
find_cdna_polya(file_path = "a")
}
\arguments{
\item{file_path}{Path of the FAST5 file}
}
\description{
Find Poly(A) tail in a cDNA read
}
