% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-cdna-polya-tail-per-read.R
\name{find_cdna_polya_tail_per_read}
\alias{find_cdna_polya_tail_per_read}
\title{Find Poly(A) tail in a single cDNA read. The read must be a poly(A) read.}
\usage{
find_cdna_polya_tail_per_read(file_path, save_plots = FALSE,
  show_plots = FALSE, save_dir = "~")
}
\arguments{
\item{file_path}{Path of the FAST5 file}
}
\value{
A list of Fast5 file data
}
\description{
Find Poly(A) tail in a single cDNA read. The read must be a poly(A) read.
}
\examples{
find_read_tail_cdna_polya('path/to/fast5/file')
}
