% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DELETE-LATER-find-cdna-polya-tails-batch-parallel.R
\name{find_cdna_polya_tails_batch_parallel}
\alias{find_cdna_polya_tails_batch_parallel}
\title{Find Poly(A) tails in FAST5 reads}
\usage{
find_cdna_polya_tails_batch_parallel(fast5_files_list, save_dir,
  csv_file_name, save_plots = FALSE, show_plots = FALSE,
  num_cores = 1)
}
\arguments{
\item{fast5_files_list}{a character list.}

\item{save_dir}{a character string. Path of the directory in which to save the
results csv file and plots (optional)}

\item{csv_file_name}{a character string. Name of the csv file in which to store the results.}

\item{save_plots}{A logical [FALSE]. Set it to TRUE if you want to
save a plot of poly(A) tail in each read, otherwise set it to FALSE.}

\item{show_plots}{A logical [FALSE]. Set it to TRUE if you want to display a
plot of poly(A) tails in each read, otherwise set it to FALSE. Setting this
option to TRUE could cause the algorithm to run very slowly.}

\item{num_cores}{}
}
\value{
A dataframe containing information about poly(A) tail lengths of each read
}
\description{
This function gets a directory of forward-strand FAST5 reads, searches
for poly(A) tails in these reads, and then returns a dataframe that contains
the data about poly-(A) tails found
}
\examples{
df <- find_cdna_polya_tails_batch_parallel('/FORWARD/STRAND/FAST5/FILES/DIRECTORY', 'SAVE/DIR', 'polya-tail-data.csv')
}
