% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-cdna-polyt-tail-per-read.R
\name{find_cdna_polyt_tail_per_read}
\alias{find_cdna_polyt_tail_per_read}
\title{Find Poly(T) tail in a single cDNA read. The read must be a poly(A) read.}
\usage{
find_cdna_polyt_tail_per_read(file_path,
  poly_a_adaptor = "GAAGATAGAGCGACAGGCAAGT", save_plots = FALSE,
  show_plots = FALSE, save_dir = "~")
}
\arguments{
\item{file_path}{Path of the FAST5 file}
}
\value{
A list of Fast5 file data
}
\description{
Find Poly(T) tail in a single cDNA read. The read must be a poly(A) read.
}
\examples{
find_cdna_polyt_tail_per_read('path/to/fast5/file')
}
