% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-cdna-tails-main-foreach.R
\name{find_cdna_tails_foreach}
\alias{find_cdna_tails_foreach}
\title{Title}
\usage{
find_cdna_tails_foreach(fast5_dir, alignment_bam_file, tails = "both",
  poly_a_adaptor = "GAAGATAGAGCGACAGGCAAGT", save_dir,
  poly_a_csv_file_name, poly_t_csv_file_name, save_plots = FALSE,
  num_cores = 1)
}
\arguments{
\item{save_plots}{}
}
\description{
Title
}
