% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-dna-tails-foreach.R
\name{find_dna_tails}
\alias{find_dna_tails}
\title{Title}
\usage{
find_dna_tails(fast5_dir, data = "cdna", save_dir = "~",
  csv_filename = "dna_tail.csv", num_cores = 1, save_plots = FALSE,
  show_plots = FALSE, plot_debug = FALSE,
  plotting_library = "rbokeh")
}
\arguments{
\item{plotting_library}{}
}
\description{
Title
}
