% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dna_tails_parallel.R
\name{find_dna_tails_parallel}
\alias{find_dna_tails_parallel}
\title{Title}
\usage{
find_dna_tails_parallel(fast5_dir, data = "cdna", save_dir = NA,
  csv_filename = "dna_tails.csv", num_cores = 1, save_plots = F,
  show_plots = F, plot_debug = F, multifast5 = F,
  basecalled_with_flipflop = F, plotting_library = "rbokeh")
}
\arguments{
\item{basecalled_with_flipflop}{}
}
\description{
Title
}
