% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/find-non-polya-region-length-without-homopolymers.R
\name{find_non_polya_region_length_without_homopolymers}
\alias{find_non_polya_region_length_without_homopolymers}
\title{Title}
\usage{
find_non_polya_region_length_without_homopolymers(non_poly_a_seq_start,
  non_poly_a_seq_end, moves_sample_wise_vector, homopolymer_threshold)
}
\arguments{
\item{homopolymer_threshold}{}
}
\value{

}
\description{
Title
}
