% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-read_tail-cdna-polya.R
\name{find_read_tail_cdna_polya}
\alias{find_read_tail_cdna_polya}
\title{Find Poly(A) tail in a single cDNA read. The read must be a poly(A) read.}
\usage{
find_read_tail_cdna_polya(file_path, generate_plots = FALSE,
  plots_directory = "~")
}
\arguments{
\item{file_path}{Path of the FAST5 file}
}
\value{
A list of Fast5 file data
}
\description{
Find Poly(A) tail in a single cDNA read. The read must be a poly(A) read.
}
\examples{
find_read_tail_cdna_polya('path/to/fast5/file')
}
