% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-rna-polya-crude-start-end.R
\name{find_rna_polya_crude_start_end}
\alias{find_rna_polya_crude_start_end}
\title{Find the crude start and end of polyA tail in RNA read}
\usage{
find_rna_polya_crude_start_end(rle_lengths, rle_values, rle_indices,
  trough_threshold)
}
\arguments{
\item{trough_threshold}{}
}
\value{

}
\description{
Find the crude start and end of polyA tail in RNA read
}
