% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-rna-polya-tails-foreach.R
\name{find_rna_polya_tails_foreach}
\alias{find_rna_polya_tails_foreach}
\title{Title}
\usage{
find_rna_polya_tails_foreach(fast5_files_list, save_dir,
  save_plots = FALSE, show_plots = FALSE, num_cores = 1,
  plotting_library = "ggplot2", plot_debug = FALSE)
}
\arguments{
\item{plot_debug}{}
}
\description{
Title
}
