% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-rna-tails-main-foreach.R
\name{find_rna_tails_foreach}
\alias{find_rna_tails_foreach}
\title{Title}
\usage{
find_rna_tails_foreach(fast5_dir, save_dir, poly_a_csv_file_name,
  save_plots = FALSE, show_plots = FALSE, plot_debug = FALSE,
  plotting_library = "ggplot2", num_cores = 1)
}
\arguments{
\item{num_cores}{}
}
\description{
Title
}
