% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-sample-index-for-fastq-base.R
\name{find_sample_index_for_fastq_base}
\alias{find_sample_index_for_fastq_base}
\title{Title}
\usage{
find_sample_index_for_fastq_base(event_data, fastq_base_number, read_type)
}
\arguments{
\item{read_type}{}
}
\value{

}
\description{
Title
}
