% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fast5-read-ids-mclapply-hdf5r.R
\name{get_fast5_read_ids_mclapply_hdf5r}
\alias{get_fast5_read_ids_mclapply_hdf5r}
\title{Title}
\usage{
get_fast5_read_ids_mclapply_hdf5r(fast5_dir, num_cores = 1)
}
\arguments{
\item{num_cores}{}
}
\description{
Title
}
