% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left-to-right-sliding-window-cdna-polyt.R
\name{left_to_right_sliding_window_cdna_polyt}
\alias{left_to_right_sliding_window_cdna_polyt}
\title{Sliding window from left to right}
\usage{
left_to_right_sliding_window_cdna_polyt(FUN, data, window_size, step_size)
}
\arguments{
\item{FUN}{Function to apply to each window's data}

\item{data}{Timeseries data to which to apply the sliding window}

\item{window_size}{Size of the window}

\item{step_size}{Step-size of the sliding window}
}
\value{
A signal that has been smoothed using the sliding window. N.B. The
  singal is now shorter in lenght
}
\description{
Applies a function to window that can slide across a timeseries from right to
left, i.e. from a read's end to its beginning.
Acknowledgment: evobiR tool
}
\examples{
left_to_right_sliding_window_cdna_polyt('mean', c(1,2,3,4,5,6,7,8), 2, 1)

}
