% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish-dna-tail-boundries.R
\name{polish_dna_tail_boundries}
\alias{polish_dna_tail_boundries}
\title{Title}
\usage{
polish_dna_tail_boundries(event_data, left_boundary, right_boundary,
  read_type)
}
\arguments{
\item{right_boundary}{}
}
\description{
Title
}
