% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-cluster-read-segments.R
\name{rca_cluster_read_segments}
\alias{rca_cluster_read_segments}
\title{Cluster segments found in a long RCA reads}
\usage{
rca_cluster_read_segments(polyat_df)
}
\arguments{
\item{polyat_df}{a dataframe as produced by the
`rca_locate_polya_polyt_segments` function}
}
\description{
Sometimes two or more completely different transcripts may be found in a
read. Multiple copies of those then need to be clustered together.
}
