% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-create-consensus-sequence.R
\name{rca_create_consensus_sequence}
\alias{rca_create_consensus_sequence}
\title{Create consensus sequence from multiple sequence}
\usage{
rca_create_consensus_sequence(data)
}
\arguments{
\item{data}{}
}
\description{
As RCA read contains multiple copies of a transcript isoform, this function
create a consensus sequence
}
