% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-find-tails-main.R
\name{rca_find_tails_main}
\alias{rca_find_tails_main}
\title{Title}
\usage{
rca_find_tails_main(file_path, read_id_fast5_file = NA,
  basecall_group = "Basecall_1D_000", save_plots = save_plots,
  show_plots = show_plots, plot_debug = FALSE, save_dir = save_dir,
  plotting_library = plotting_library, multifast5 = multifast5,
  basecalled_with = basecalled_with, model = model, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
