% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-find-tails-per-read.R
\name{rca_find_tails_per_read}
\alias{rca_find_tails_per_read}
\title{Title}
\usage{
rca_find_tails_per_read(rca_data, read_data, file_path)
}
\arguments{
\item{file_path}{}
}
\description{
Title
}
