% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-get-segment-boundaries-in-samples.R
\name{rca_get_segment_boundaries_in_samples}
\alias{rca_get_segment_boundaries_in_samples}
\title{Title}
\usage{
rca_get_segment_boundaries_in_samples(event_data, fastq_start_index,
  fastq_end_index, read_type, raw_data_len, samples_per_nt)
}
\arguments{
\item{samples_per_nt}{}
}
\description{
Title
}
