% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-locate-polya-polyt-segments.R
\name{rca_locate_polya_polyt_segments}
\alias{rca_locate_polya_polyt_segments}
\title{Find polyA/PolyT segments within the long RCA reads}
\usage{
rca_locate_polya_polyt_segments(ot_rc_ot_df, fastq)
}
\arguments{
\item{ot_rc_ot_df}{a dataframe as returned by the rca_search_oligo function}

\item{fastq}{a DNA string. The fastQ sequence of the read}
}
\description{
Find polyA/PolyT segments within the long RCA reads
}
