% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-make-slope-and-mean-signals.R
\name{rca_make_slope_and_mean_signals}
\alias{rca_make_slope_and_mean_signals}
\title{Title}
\usage{
rca_make_slope_and_mean_signals(data, window_size, step_size)
}
\arguments{
\item{step_size}{}
}
\description{
Title
}
