% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-search-oligos.R
\name{rca_search_oligo}
\alias{rca_search_oligo}
\title{Search for OligoT and reverse complement oligoT}
\usage{
rca_search_oligo(fastq, ot, rc_ot, gfp, rc_gfp, ssw_before, rc_ssw_before,
  ssw_after, rc_ssw_after, submat)
}
\arguments{
\item{submat}{}
}
\value{
a dataframe containing `start`, `stop` index of oligo in the FastQ
and `what_is_it` that was found (`ot` or `rc_or`). A NULL is returned if
no oocurance of any oligo is found.
}
\description{
Search for OligoT and reverse complement oligoT
}
