% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right-to-left-sliding-window-cdna-polya.R
\name{right_to_left_sliding_window_cdna_polya}
\alias{right_to_left_sliding_window_cdna_polya}
\title{Sliding window from right to left.}
\usage{
right_to_left_sliding_window_cdna_polya(FUN, data, window_size, step_size)
}
\arguments{
\item{FUN}{Function to apply to each window's data}

\item{data}{Timeseries data to which to apply the sliding window}

\item{window_size}{Size of the window}

\item{step_size}{Step-size of the sliding window}
}
\value{

}
\description{
Applies a function to window that can slide across a timeseries from left to
right of a cDNA poly(A) read, i.e. from a read's beginning to its end.
}
