% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding-windows-rna.R
\name{right_to_left_sliding_window_rna}
\alias{right_to_left_sliding_window_rna}
\title{Sliding window from right to left.}
\usage{
right_to_left_sliding_window_rna(FUN, data, window_size, step_size)
}
\arguments{
\item{FUN}{Function to apply to each window's data}

\item{data}{Timeseries data to which to apply the sliding window}

\item{window_size}{Size of the window}

\item{step_size}{Step-size of the sliding window}
}
\value{
A signal that has been smoothed using the sliding window.
}
\description{
Applies a function to window that can slide across a timeseries from right to
left i.e., from a read's end to its begining.
Acknowledgment: evobiR tool
}
