% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding-window.R
\name{sliding_window}
\alias{sliding_window}
\title{Sliding window}
\usage{
sliding_window(FUN, data, window_size, step_size)
}
\arguments{
\item{FUN}{Function to apply to each window's data}

\item{data}{Timeseries data to which to apply the sliding window}

\item{window_size}{Size of the window}

\item{step_size}{Step-size of the sliding window}
}
\value{
A signal that has been smoothed using the sliding window. N.B. The
  singal is now shorter in lenght
}
\description{
Applies a function to window that can slide across a timeseries
Acknowledgment: evobiR tool
}
\examples{
sliding_window('mean', c(1,2,3,4,5,6,7,8), 2, 1)

}
