% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothen-rle-intersections-rna.R
\name{smoothen_rle_intersections_rna}
\alias{smoothen_rle_intersections_rna}
\title{Smoothen RLE intersections for RNA}
\usage{
smoothen_rle_intersections_rna(rle_intersections)
}
\arguments{
\item{rle_intersections}{RLE object}
}
\value{

}
\description{
Smoothen RLE intersections for RNA
}
