% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-normalize.R
\name{z_norm}
\alias{z_norm}
\title{z-normalize}
\usage{
z_norm(timeseries)
}
\arguments{
\item{timeseries}{Timeseries data that needs to be z-normalized}
}
\value{
z_normalized_timeseries A z-normalized timeseries
}
\description{
Apply z-normalization to the time series data
}
\examples{
z_norm(c(1,2,3,4,5))
}
