% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-rna-polya-precise-start-end.R
\name{find_rna_polya_precise_start_end}
\alias{find_rna_polya_precise_start_end}
\title{Title}
\usage{
find_rna_polya_precise_start_end(truncated_data, POLY_A_RNA_THRESHOLD,
  crude_polya_boundries, save_plots, show_plots)
}
\arguments{
\item{crude_polya_boundries}{}
}
\value{

}
\description{
Title
}
