% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rna_polya_tail_per_read.R
\name{find_rna_polya_tail_per_read}
\alias{find_rna_polya_tail_per_read}
\title{Find Poly(A) tail in a single RNA read.}
\usage{
find_rna_polya_tail_per_read(file_path, save_plots = FALSE,
  show_plots = FALSE, save_dir = "~", plotting_library = "rbokeh",
  plot_debug = FALSE)
}
\arguments{
\item{file_path}{Path of the FAST5 file}
}
\value{
A list of Fast5 file data
}
\description{
Find Poly(A) tail in a single RNA read.
}
\examples{
find_rna_polya_tail_per_read('path/to/fast5/file')
}
