% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sample_index_for_fastq_base.R
\name{find_sample_index_for_fastq_base}
\alias{find_sample_index_for_fastq_base}
\title{Finds raw data sample index corresponding to a base.}
\usage{
find_sample_index_for_fastq_base(event_data, fastq_base_number, read_type)
}
\arguments{
\item{read_type}{}
}
\value{

}
\description{
Finds raw data sample index in the event data table corresponding to a FastQ
base number. The closet possible matching sample is dependent on the type of
the tail.
}
