% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca-search-oligos.R
\name{rca_search_oligo}
\alias{rca_search_oligo}
\title{Search for OligoT and reverse complement oligoT}
\usage{
rca_search_oligo(fastq, ot, rc_ot, submat)
}
\arguments{
\item{fastq}{a string. The fastQ sequence of the read.}

\item{ot}{a DNA string. Oligo T sequence (5' to 3')}

\item{rc_ot}{a DNA string. Reverese complement Oligo T sequence (5' to 3')}

\item{submat}{A Biostring substitution matrix}
}
\value{
a dataframe containing `start`, `stop` index of oligo in the FastQ
and `what_is_it` that was found (`ot` or `rc_or`). A NULL is returned if
no oocurance of any oligo is found.
}
\description{
Search for OligoT and reverse complement oligoT
}
